feat(历史记录): 添加支持按照年月日以及所有的删除历史记录功能
日历上添加右键菜单；
调整 segment_id 的数据类型

+++ WindowsFormsApplication2/History/DataType.cs
@@ -30,6 +30,6 @@ namespace WindowsFormsApplication2.History
         /// <summary>
         /// 文段 id
         /// </summary>
-        public int SegmentId { get; set; }
+        public long SegmentId { get; set; }
     }
 }
